+++ CHANGELOG.md
@@ -11,6 +11,12 @@ From v2.0.0 to v2.0.1
 - Papyrus descriptors are not fetched by default anymore from the `Papyrus`  adapter, which caused fetching of unnecessary data.
 
 ## Changes
+- `PCMSplit` replaces `StratifiedPerTarget` and is compatible with `RandomSplit`, `ScaffoldSplit` and `ClusterSplit`.
+- In the case single-task dataset, the `RandomSplit` now uses `StratifiedShuffleSplit` in case of classification.
 
 ## New Features
+- `ClusterSplit` - splits data based clustering of molecular fingerprints.
 - raise error if search space for optuna optimization is missing search space type annotation or if type not in list
+
+## Removed Features
+- `StratifiedPerTarget` is replaced by `PCMSplit`.